x/requirements.txt
cozmo[camera]
Pillow<=9.4.0
pandas
numpy
opencv-python
matplotlib
selenium
webdriver-manager